app/requirements.txt
# Requirements for LangChain Multi-LLM project

# Core LangChain
langchain>=0.1.0
langchain-core>=0.1.0

# LLM Integrations
langchain-google-genai>=1.0.0
langchain-anthropic>=0.1.0

# Utilities
python-dotenv>=1.0.0

# Optional: for advanced features
langchain-community>=0.0.20

# OpenTelemetry for observability
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-exporter-otlp-proto-grpc>=1.21.0
opentelemetry-instrumentation-requests>=0.42b0

# Streamlit for UI
streamlit>=1.31.0
